+++ R/expand_model_out_grid.R
@@ -36,6 +36,11 @@
 #' character which can be faster when large expanded grids are expected.
 #' If `required_vals_only = TRUE`, values are limited to the combinations of required
 #' values only.
+#'
+#' Note that if `required_vals_only = TRUE` and an optional output type is
+#' requested through `output_types`, a zero row grid will be returned.
+#' If all output types are requested however (i.e. when `output_types = NULL`) and
+#' they are all optional, a grid of required task ID values only will be returned.
 #' @inheritParams hubData::coerce_to_hub_schema
 #' @details
 #' When a round is set to `round_id_from_variable: true`,
@@ -166,6 +171,9 @@ expand_model_out_grid <- function(config_tasks,
     config_tasks, round_id
   )
   round_config <- get_round_config(config_tasks, round_id)
+  # Create a logical variable to control what is returned by expand_output_type_grid.
+  # See not in fn for details.
+  all_output_types <- is.null(output_types) # nolint: object_usage_linter
 
   task_id_l <- purrr::map(
     round_config[["model_tasks"]],
@@ -211,7 +219,8 @@ expand_model_out_grid <- function(config_tasks,
     task_id_l, output_type_l,
     ~ expand_output_type_grid(
       task_id_values = .x,
-      output_type_values = .y
+      output_type_values = .y,
+      all_output_types = all_output_types
     )
   )
 
@@ -245,7 +254,27 @@ process_grid_inputs <- function(x, required_vals_only = FALSE) {
 # Function that expands modeling task level lists of task IDs and output type
 # values into a grid and combines them into a single tibble.
 expand_output_type_grid <- function(task_id_values,
-                                    output_type_values) {
+                                    output_type_values,
+                                    all_output_types = TRUE) {
+  # Return a grid of only task IDs if no output type values are provided but
+  # only if a specific output type subset is not requested.
+  # Otherwise return a zero row grid.
+  # No output type values can either be the result of all optional output types
+  # when required values only are requested or as a result of output type sub-setting.
+  # When requesting required values only for an entire rounds (i.e. all output types),
+  # we want required task ID values to be returned, even is all output types are
+  # optional. However, it does not make sense to return required values for an
+  # optional output type when a user is specifically requesting required values
+  # for an output type. In that situation it's more appropriate to return a zero row grid.
+  if (length(output_type_values) == 0 && all_output_types) {
+    return(
+      expand.grid(
+        purrr::compact(task_id_values),
+        stringsAsFactors = FALSE
+      )
+    )
+  }
+
   purrr::imap(
     output_type_values,
     ~ c(task_id_values, list(