+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development version)
+
 # hubValidations 0.6.2
 
 * Fixed bug in `check_tbl_col_types()` where columns in model output data with more than one class were causing an EXEC error (#118). Thanks for the bug report @ruarai!