+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development version)
+
 # hubValidations 0.7.0
 
 * Added function `create_custom_check()` for creating custom validation check function files from templates (#121).